+++ cmd/admission/main.go
@@ -45,8 +45,8 @@ func registerPlugins(app *app.App) error {
 	admissionGpuSharingPlugin := gpusharing.New(app.Client, app.Options.GPUSharingEnabled)
 	admissionPlugins.RegisterPlugin(admissionGpuSharingPlugin)
 
-	if app.Options.GPUPodRuntimeClassName != "" {
-		admissionRuntimeEnforcementPlugin := runtimeenforcement.New(app.Options.GPUPodRuntimeClassName)
+	if rc := app.Options.ResolvedGPUFractionRuntimeClassName(); rc != "" {
+		admissionRuntimeEnforcementPlugin := runtimeenforcement.New(rc)
 		admissionPlugins.RegisterPlugin(admissionRuntimeEnforcementPlugin)
 	}