+++ pkg/binder/binding/resourcereservation/resource_reservation.go
@@ -56,7 +56,6 @@ type service struct {
 	serviceAccountName                  string
 	appLabelValue                       string
 	scalingPodNamespace                 string
-	runtimeClassName                    string
 	podResources                        *v1.ResourceRequirements
 	reservationPodSecurityContext       *v1.PodSecurityContext
 	reservationContainerSecurityContext *v1.SecurityContext
@@ -71,7 +70,6 @@ func NewService(
 	serviceAccountName string,
 	appLabelValue string,
 	scalingPodNamespace string,
-	runtimeClassName string,
 	podResources *v1.ResourceRequirements,
 	reservationPodSecurityContext *v1.PodSecurityContext,
 	reservationContainerSecurityContext *v1.SecurityContext,
@@ -86,7 +84,6 @@ func NewService(
 		serviceAccountName:                  serviceAccountName,
 		appLabelValue:                       appLabelValue,
 		scalingPodNamespace:                 scalingPodNamespace,
-		runtimeClassName:                    runtimeClassName,
 		podResources:                        podResources,
 		reservationPodSecurityContext:       reservationPodSecurityContext,
 		reservationContainerSecurityContext: reservationContainerSecurityContext,
@@ -531,13 +528,7 @@ func (rsc *service) createResourceReservationPod(
 			},
 		},
 		Spec: v1.PodSpec{
-			NodeName: nodeName,
-			RuntimeClassName: func() *string {
-				if len(rsc.runtimeClassName) == 0 {
-					return nil
-				}
-				return &rsc.runtimeClassName
-			}(),
+			NodeName:           nodeName,
 			SecurityContext:    rsc.reservationPodSecurityContext,
 			ServiceAccountName: rsc.serviceAccountName,
 			Containers: []v1.Container{